fix: loading and filter proudcts error
