+++ packages/vanilla/src/utils/productUtils.js
@@ -1,0 +1,50 @@
+export function filterProducts(products, query) {
+  let filtered = [...products];
+
+  // 검색어 필터링
+  if (query.search) {
+    const searchTerm = query.search.toLowerCase();
+    filtered = filtered.filter(
+      (item) => item.title.toLowerCase().includes(searchTerm) || item.brand.toLowerCase().includes(searchTerm),
+    );
+  }
+
+  // 카테고리 필터링
+  if (query.category1) {
+    filtered = filtered.filter((item) => item.category1 === query.category1);
+  }
+  if (query.category2) {
+    filtered = filtered.filter((item) => item.category2 === query.category2);
+  }
+
+  // 정렬
+  if (query.sort) {
+    switch (query.sort) {
+      case "price_asc":
+        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
+        break;
+      case "price_desc":
+        filtered.sort((a, b) => parseInt(b.lprice) - parseInt(a.lprice));
+        break;
+      case "name_asc":
+        filtered.sort((a, b) => a.title.localeCompare(b.title, "ko"));
+        break;
+      case "name_desc":
+        filtered.sort((a, b) => b.title.localeCompare(a.title, "ko"));
+        break;
+      default:
+        // 기본은 가격 낮은 순
+        filtered.sort((a, b) => parseInt(a.lprice) - parseInt(b.lprice));
+    }
+  }
+
+  return filtered;
+}
+
+export function loadingProudcts(products) {
+  const byPriceAsc = (a, b) => Number(a.lprice || 0) - Number(b.lprice || 0);
+  // 정렬 적용 (기본: price_asc)
+  let sorted = [...products];
+
+  return sorted.sort(byPriceAsc);
+}
